feat: scheduler date switching
