+++ frontend/src/pages/Scheduler/Scheduler.jsx
@@ -4,38 +4,20 @@ import { setPageTitle } from '/src/actions'
 
 import Calendar from '/src/containers/Calendar'
 import SchedulerNav from './SchedulerNav'
-
-const getWeekStart = () => {
-  const now = new Date()
-  const dayOfWeek = now.getDay()
-  const diff = now.getDate() - dayOfWeek + (dayOfWeek === 0 ? -6 : 1)
-  const weekStart = new Date(now.setDate(diff))
-  weekStart.setHours(0, 0, 0, 0)
-  return weekStart
-}
+import { getWeekStart, createTitle } from './utils'
 
 const Scheduler = () => {
   const dispatch = useDispatch()
-  const [startDate, setStartDate] = useState(getWeekStart())
-
-  const pageTitle = useMemo(() => {
-    const start = startDate.toLocaleDateString('en-US', {
-      month: 'short',
-      day: 'numeric',
-    })
-    const end = new Date(
-      startDate.getTime() + 6 * 24 * 60 * 60 * 1000
-    ).toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
-    return `Scheduler (${start} - ${end})`
-  }, [startDate])
+  const [startTime, setStartTime] = useState(getWeekStart())
 
   useEffect(() => {
+    const pageTitle = createTitle(startTime)
     dispatch(setPageTitle({ title: pageTitle }))
-  }, [])
+  }, [startTime])
 
   return (
     <main className="column">
-      <SchedulerNav />
+      <SchedulerNav startTime={startTime} setStartTime={setStartTime} />
       <section className="grow nopad">
         <Calendar />
       </section>
